% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Demo NOS dataset}
\format{
A data frame with 8 studies and NOS domains
}
\usage{
demo_data
}
\description{
Example dataset for Newcastle-Ottawa Scale visualization.
}
\keyword{datasets}
